requirements.txt
streamlit
uvicorn
fastapi
python-dotenv
langchain_groq
langgraph
ipython
langchain_community
feedparser
sentence_transformers
pytest
deepeval
openai
langsmith
einops
guardrails-ai
guardrails-grhub-ban-list
guardrails-grhub-toxic-language
torch
torchvision
torchaudio